requirements.txt
pypdf
langchain
google-generativeai
streamlit
python-dotenv